This is a clone of Rejouice.com using HTML, CSS and javaScript
This project is a clone of the Rejouice website built using HTML, CSS, and JavaScript. Rejouice is a platform for sharing joyful moments with friends and family. With this clone, users can experience the key features of Rejouice, including creating and sharing joyful memories, connecting with loved ones, and spreading positivity.